moved udo-related code to udo.c
