+++ H/udo.h
@@ -65,6 +65,8 @@ typedef struct {
     MYFLT   *args[OPCODENUMOUTS_MAX];
 } XOUT_MAX;
 
+int32_t useropcd(CSOUND *, UOPCODE *p);
+int32_t useropcdset(CSOUND *, UOPCODE *p);
 int32_t useropcd_local_ksmps(CSOUND *, UOPCODE*);
 int32_t useropcd_pass_by_copy(CSOUND *, UOPCODE*);
 int32_t useropcd_pass_by_ref(CSOUND *, UOPCODE *);
